Revert "Fb3 progress, please reject this pull request"
