+++ flamerobin_flamerobin.vcxproj.filters
@@ -333,9 +333,6 @@
     <ClCompile Include="src\metadata\view.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\gui\InsertParametersDialog.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\gui\AboutBox.h">
@@ -654,9 +651,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\metadata\view.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="src\gui\InsertParametersDialog.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
