+++ makefile.vc
@@ -56,7 +56,7 @@ TARGET_CPU = $(CPU)
 FLAMEROBIN_CXXFLAGS = /M$(__RTL_TYPE)$(__DEBUGINFO_0) /DWIN32 $(__DEBUGINFO) \
 	$(____DEBUGINFO) /Fdvcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin.pdb \
 	$(__OPTIMIZE) /W4 $(____flamerobin_WXDEBUG_p) $(____flamerobin_DEBUGFLAG_p) \
-	/D_WINDOWS /D__WINDOWS__ /DWINVER=0x500 /DWIN32 /D__WIN32__ /D__WIN95__ \
+	/D_WINDOWS /D__WINDOWS__ /DWINVER=0x400 /DWIN32 /D__WIN32__ /D__WIN95__ \
 	/DSTRICT /D__WXMSW__ /DwxUSE_GUI=1 /DwxUSE_REGEX=1 /DwxUSE_UNICODE=1 \
 	/DWIN32_LEAN_AND_MEAN \
 	/I$(____flamerobin_WX_LIB_DIR_FILENAMES)\mswu$(D_OPT) \
@@ -340,7 +340,7 @@ revision-info:
 
 vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin.exe: vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin_frprec.obj  $(FLAMEROBIN_OBJECTS) vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp.lib vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin_flamerobin.res vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp.lib
 	link /NOLOGO /OUT:$@   $(__DEBUGINFO_5) /pdb:"vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin.pdb" $(__DEBUGINFO_9) /nologo /subsystem:windows $(LINK_TARGET_CPU) $(______flamerobin_WX_LIB_DIR_FILENAMES_2_p) /LIBPATH:$(BOOST_LIB_DIR) /SUBSYSTEM:WINDOWS $(LDFLAGS) @<<
-	$(FLAMEROBIN_OBJECTS) $(FLAMEROBIN_RESOURCES)  wxmsw31u$(D_OPT)_aui.lib wxmsw31u$(D_OPT)_stc.lib wxmsw31u$(D_OPT)_html.lib wxmsw31u$(D_OPT)_adv.lib wxmsw31u$(D_OPT)_core.lib wxbase31u$(D_OPT)_xml.lib wxbase31u$(D_OPT).lib wxexpat$(D_OPT).lib wxtiff$(D_OPT).lib wxjpeg$(D_OPT).lib wxpng$(D_OPT).lib wxzlib$(D_OPT).lib wxregexu$(D_OPT).lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib comctl32.lib rpcrt4.lib wsock32.lib vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp.lib
+	$(FLAMEROBIN_OBJECTS) $(FLAMEROBIN_RESOURCES)  wxmsw30u$(D_OPT)_aui.lib wxmsw30u$(D_OPT)_stc.lib wxmsw30u$(D_OPT)_html.lib wxmsw30u$(D_OPT)_adv.lib wxmsw30u$(D_OPT)_core.lib wxbase30u$(D_OPT)_xml.lib wxbase30u$(D_OPT).lib wxexpat$(D_OPT).lib wxtiff$(D_OPT).lib wxjpeg$(D_OPT).lib wxpng$(D_OPT).lib wxzlib$(D_OPT).lib wxregexu$(D_OPT).lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib comctl32.lib rpcrt4.lib wsock32.lib vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp.lib
 <<
 
 vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp.lib: vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp__ibpp.obj  $(IBPP_OBJECTS)
@@ -668,7 +668,7 @@ vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin_StyleGuideMSW.obj: .\src\gui\msw
 	$(CXX) /c /nologo /TP /Fo$@ $(FLAMEROBIN_CXXFLAGS) .\src\gui\msw\StyleGuideMSW.cpp
 
 vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\flamerobin_flamerobin.res: .\res\flamerobin.rc
-	rc /fo$@  /d WIN32 $(____DEBUGINFO_6_p_1) $(____flamerobin_WXDEBUG_p_1) $(____flamerobin_DEBUGFLAG_p_1) /d _WINDOWS /d __WINDOWS__ /d WINVER=0x500 /d WIN32 /d __WIN32__ /d __WIN95__ /d STRICT /d __WXMSW__ /d wxUSE_GUI=1 /d wxUSE_REGEX=1 /d wxUSE_UNICODE=1 /d WIN32_LEAN_AND_MEAN /i $(____flamerobin_WX_LIB_DIR_FILENAMES_1)\mswu$(D_OPT) /i $(WXDIR)\contrib\include /i $(WXDIR)\include /i $(BOOST_ROOT) /d _WINDOWS /d IBPP_WINDOWS /i . /i .\src /i .\src\ibpp /i .\res .\res\flamerobin.rc
+	rc /fo$@  /d WIN32 $(____DEBUGINFO_6_p_1) $(____flamerobin_WXDEBUG_p_1) $(____flamerobin_DEBUGFLAG_p_1) /d _WINDOWS /d __WINDOWS__ /d WINVER=0x400 /d WIN32 /d __WIN32__ /d __WIN95__ /d STRICT /d __WXMSW__ /d wxUSE_GUI=1 /d wxUSE_REGEX=1 /d wxUSE_UNICODE=1 /d WIN32_LEAN_AND_MEAN /i $(____flamerobin_WX_LIB_DIR_FILENAMES_1)\mswu$(D_OPT) /i $(WXDIR)\contrib\include /i $(WXDIR)\include /i $(BOOST_ROOT) /d _WINDOWS /d IBPP_WINDOWS /i . /i .\src /i .\src\ibpp /i .\res .\res\flamerobin.rc
 
 vcu$(R_OPT)$(D_OPT)$(DIR_SUFFIX_CPU)\ibpp__dpb.obj: .\src\ibpp\_dpb.cpp
 	$(CXX) /c /nologo /TP /Fo$@ $(IBPP_CXXFLAGS) .\src\ibpp\_dpb.cpp
