+++ src/gui/EditBlobDialog.cpp
@@ -127,12 +127,12 @@ void EditBlobDialogSTC::setIsNull(bool isNull)
     {
         wxStyledTextCtrl::SetText("[null]");
         SelectAll();
-        StartStyling(0, 0);
+        StartStyling(0);
         SetStyling(GetTextLength(), 0x0A);
     }
     else
     {
-        StartStyling(0, 0);
+        StartStyling(0);
         SetStyling(GetTextLength(), 0);
     }
     isNullM = isNull;
@@ -786,7 +786,7 @@ bool EditBlobDialog::loadFromStreamAsBinary(wxInputStream& stream, bool isNull, 
         }
     }
     // Set text styling
-    blob_binary->StartStyling(0, 0);
+    blob_binary->StartStyling(0);
     for (int i = 0; i < blob_binary->GetLineCount(); i++)
         blob_binary->SetStyleBytes(CharsPerLine, &styleBytes[0]);
     progressEnd();
